requirements.txt
astroid==2.12.13
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
dill==0.3.6
fake-useragent==1.1.0
flake8==6.0.0
idna==3.4
isort==5.10.1
lazy-object-proxy==1.8.0
lxml==4.9.1
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
platformdirs==2.5.4
pycodestyle==2.10.0
pydantic==1.10.2
pyflakes==3.0.1
pylint==2.15.7
pytest==7.2.0
python-decouple==3.6
pyuser-agent==0.1.0
requests==2.28.1
schedule==1.1.0
soupsieve==2.3.2.post1
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
urllib3==1.26.13
wrapt==1.14.1
